Deleted .vim/ and .tmux/ folders.
They contain plugin's git repositories, which become embedded into my
repo. Cloning my repo will not get the contents of these folders,
either.

The main reason I wanted to put these folders in my repo was the plugin
managers. With this I wanted to automate "installing" of plugin managers
in vim and tmux. But this will not be possible, because of the point I
made in previous paragraph.